+++ speech2text/engines/ds.py
@@ -1,6 +1,8 @@
 from speech2text.engines import StreamingSTT, StreamThread, STT
 import requests
 from queue import Queue
+from speech2text.log import LOG
+from os.path import isfile
 
 
 class DeepSpeechServerSTT(STT):
@@ -53,4 +55,56 @@ class DeepSpeechStreamServerSTT(StreamingSTT):
             self.lang,
             self.config.get('stream_uri')
         )
+
+
+class DeepSpeechSTT(STT):
+
+    def __init__(self, config):
+        super().__init__(config)
+        global DeepSpeechModel, np
+        from deepspeech import Model as DeepSpeechModel
+        import numpy as np
+        model = self.config.get("model")
+        scorer = self.config.get("scorer")
+        if not model or not isfile(model):
+            LOG.error("You need to provide a valid model file")
+            LOG.info("download a model from https://github.com/mozilla/DeepSpeech")
+            raise FileNotFoundError
+        if not scorer or not isfile(scorer):
+            LOG.warning("You should provide a valid scorer")
+            LOG.info("download scorer from https://github.com/mozilla/DeepSpeech")
+        self.ds = DeepSpeechModel(model)
+        if scorer:
+            self.ds.enableExternalScorer(scorer)
+
+    def execute(self, audio, language=None):
+        audio = np.frombuffer(audio.frame_data, np.int16)
+        return self.ds.stt(audio)
+
+
+class DeepSpeechStreamThread(StreamThread):
+    def __init__(self, queue, language, ds):
+        super().__init__(queue, language)
+        self.ds = ds
+
+    def handle_audio_stream(self, audio, language):
+        for a in audio:
+            data = np.frombuffer(a, np.int16)
+            self.ds.feedAudioContent(data)
+            # self.text = self.ds.intermediateDecode()
+        return self.text
+
+    def finalize(self):
+        self.text = self.ds.finishStream()
+
+
+class DeepSpeechStreamSTT(StreamingSTT, DeepSpeechSTT):
+    def create_streaming_thread(self):
+        self.queue = Queue()
+        return DeepSpeechStreamThread(
+            self.queue,
+            self.lang,
+            self.ds.createStream()
+        )
+